+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 券種グループサービスをオファーカテゴリーサービスへ移行
+- subjectサービスをaccountTitleサービスへ移行
 
 ### Deprecated